opt.: tool view Fixes #50
