+++ lib/view/page/home/home.dart
@@ -17,6 +17,7 @@ import 'package:gpt_box/core/util/image.dart';
 import 'package:gpt_box/core/util/sync/base.dart';
 import 'package:gpt_box/data/model/chat/config.dart';
 import 'package:gpt_box/data/model/chat/history.dart';
+import 'package:gpt_box/data/model/chat/history.ext.dart';
 import 'package:gpt_box/data/model/chat/type.dart';
 import 'package:gpt_box/data/provider/all.dart';
 import 'package:gpt_box/data/res/build.dart';
@@ -26,10 +27,6 @@ import 'package:gpt_box/data/res/rnode.dart';
 import 'package:gpt_box/data/res/url.dart';
 import 'package:gpt_box/data/store/all.dart';
 import 'package:gpt_box/view/widget/audio.dart';
-import 'package:gpt_box/view/widget/code.dart';
-import 'package:gpt_box/view/widget/image.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:flutter_markdown_latex/flutter_markdown_latex.dart';
 import 'package:icons_plus/icons_plus.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:screenshot/screenshot.dart';
